slider.ts composable created.
